Push notes to repo for others to use

+++ qemu-csd/bpf_helpers/include/bpf_helpers_prog.h
@@ -34,7 +34,7 @@
 static void (*bpf_return_data)(void *data, uint64_t size) = (void *) 1;
 
 // Called by BPF to perform an 'on device' read at a specific LBA.
-static void (*bpf_read)(uint64_t lba, uint64_t offset, uint16_t limit, void *data) = (void *) 2;
+static void (*bpf_read)(uint64_t lba, uint64_t offset, uint64_t limit, void *data) = (void *) 2;
 
 // Called by BPF to get the size in bytes per LBA.
 static uint64_t (*bpf_get_lba_size)(void) = (void *) 3;
